chore: add .gitignore, .trashconfig, LICENSE, README.md, Taskfile.yml, and main.go files
The .gitignore file is added to specify which files and directories should be ignored by Git.

The .trashconfig file is added to define the directory where the trash files will be stored.

The LICENSE file is added to specify the license under which the software is distributed.

The README.md file is added to provide information about the project, including installation instructions, usage examples, and available commands.

The Taskfile.yml file is added to define tasks for running, testing, building, and installing the project.

The main.go file is added as the entry point of the application.